# Refactor: Optimize context switching code for 68k architecture