@mikrosk 2025-06-13
Thanks for the courage to touch context.S.

However this PR screams LLM ("AI") all over the place. That itself isn't a bad thing but it definitely shows in the code. For instance that "Prefetch Optimization in change_context ⭐ Performance Improvement", really? Could you measure any speed improvement over the original code?

Also what about all that crappy comments, like "// push the frame/vector word" => "// standard frame/vector word" or explicitly mentioned that this code was optimized or that this code is for saving registers etc... that's just... not useful.

What I would appreciate much more is separate commits which each change: i.e. the FPU functions would be separated / changed one by one so it would be easy to verify which code is replaced by which.