@MedourMehdi 2025-06-13
Hi Mikro

Of course it screams LLM, in other way optimisation would be already done by you guys, and notice I'm only an asm 68k newby.

As I'm working on threads implementation I need to improve context switching speed and efficiency, so I'm actualy writing a test code in order to mesure thread switching behavior.

Whatever I read the diff for context.S and it make sense to me.

I didn't push all optimizations yet, just the fpu handling and change/save context optimisation but I'll split it in two part if you want.