@MedourMehdi 2025-06-13
I closed this PR: LLM is really a bullshit for now - You're right, my mistake :)