@mikrosk 2025-06-14
I think the FPU code unification was a good idea but it would require more granular commits. One character overlooked in such a critical function can lead to catastrophic results (I learned the hard way) so it needs to be as clean as possible.