@mikrosk 2025-06-14
Whatever you prefer but you are free to completely replace even the current branch (with force push).