@MedourMehdi 2025-06-14
This [commit](https://github.com/freemint/freemint/pull/412/commits/21b3ca082533fc2941574295a18ff589790bbada) gives me results under 13ms per context switch (4 threads/10000 ctx switch for each) -> no way with non optimized context.S