@MedourMehdi 2025-06-14
**I atleast won ~10% of perf with context switching**

<img width="432" alt="Capture d’écran 2025-06-15 à 01 36 28" src="https://github.com/user-attachments/assets/15219212-41ce-46a6-8276-872d00f28881" />