@MedourMehdi 2025-06-17
Tis was tested under Hatari 000/020/030 and Aranym(040+MMU)... All is working as it should.