@mikrosk 2025-06-20
Good job in getting some real measurements. However this PR still needs some attention.

You mentioned that "There's one commit who should show the diff with the original one" ... I still see like 11 commits in this PR, which one were you referring to?

Then there's the insane amounts of small changes which have no value whatsoever, e.g.
![image](https://github.com/user-attachments/assets/87369b28-c0c5-43e0-9111-05a8b5878054)

and while it is good that it is all in one separate commit, its decription:
```
Clarified ambiguous comments while preserving originals:
- Added "// (faster than cmp)" in build_context
- Explained "Adjust loop counter" comments
- Clarified VM page touching as "Touch page for VM"
- Explained ATC flushing in change_context
- Clarified FPU null frame checks
- Added explicit "Skip if FPU not in use"
- Noted when we're saving/restoring control vs data registers
```
doesn't really say what was your motivation for those changes. While technically correct, I don't think the comments are now _that much_ better clarified and it's hundreds and hundreds of changes. If you are changing something, please change it for yourself (and t herefore you know why did you so) instead of letting AI to do it for you for whatever reasons.

What would I propose: since you are now aware which of the optimisations helped and by how much, force-push a new branch (or create a new PR, up to you) with _as minimal changes as possible_. Therefore your PR might look like:

- commit 1: separate stuff into macros / functions, no functional changes, all compilable and working
- commit 2: optimisation X, explaining its benefit, ideally with some numbers
- commit 3: optimisation Y, explaining its benefit, ideally with some numbers
- ...

In such case you are free to change any comments YOU find unsuitable / unclear, because it is you who is making those optimisations and you know what was the reason behind them. Also, each commit will be easily reviewable, with clear intent etc.

There's no point in pushing 11 commits into master with hundreds of lines changes only to have one optimisation doing something well.