@MedourMehdi 2025-06-20
Sorry guys but I prefer close the request -  I've no real support here and comments like "let the AI" or "for your self" are not realy welcome. Moreover PR or support demands are not managed. I'll make changes in my own repo and i let you do end of life maintenance for this one.