@mikrosk 2025-06-20
No need for such a drastic measures but of course, it's your decision.

You have to understand, the resources for accepting new code are pretty limited. So any new code must:

- bring something good (ideally fixing a bug, introducing an useful feature requested by others, supporting a new hardware, ...)
- be easy to test
- be easy to review and verify that it does indeed what it promises
- be maintainable in the future, when you decide to focus on other things

While your PRs more or less adhere to the first two bullet points, the rest is not so straight-forward.

Flooding the kernel with thousands of AI generated lines of new code doesn't adhere to these simple requirements, and I don't think you'd have succeeded with them even in projects like Linux which has _much more_ people able to review your code.

So yes, unless you don't want to end up like Helmut with his infamous XaAES branch (which basically nobody used), you need to adhere to the boring and annoying contributor rules, that's life. Even I, owner of this repository, don't allow myself pushing anything I like and always go through the review process.

You are free to work on whatever you like and find enjoyable and you are more than welcome to come back with a proper / final contribution to be discussed and improved upon.