@MedourMehdi 2025-06-20
Where I spend hundreds of hours coding your talking about AI in every your comments! That's disrespect. AI just indent, comments and search for bugs - But never mind, as my formers PR didn't used any AI and were still ignored I'll prefer code for my infamous project :)