@mikrosk 2025-06-20
I didn't speak for all PRs, I spoke about this concrete one. The others I didn't even see, they were outside of my expertise. And this one did contain a lot of AI generated content. Not sure why do you take an offence in that, I wouldn't care even if a bear created that code  if it were in compliance with the traditional (and in freemint quite relaxed) contributor guidelines.

You've invested hundreds of hours into your work - that's great but you can't expect reviewers to do the same. It's your code and therefore your responsibility to make sure that your changes are easy to follow, easy to explain and discuss about. That's how open source contributing works and I have to adhere to those rules as well, in any project except my own private ones.

Anytime you reconsider, your contributions are welcome.