@MedourMehdi 2025-06-22
Yes, I understand.

But to be more efficient, I'll use my repository for now, as I'm not familiar with these development rules (like not committing a comment rework, blah blah blah...).

In any case, feel free to grab any commit you want from my repository and rework it as needed.

I simply can't wait a year or more for a trivial commit (like langinfo!); it's absurd.

So, to summarize, since there isn't enough staff or time to manage commits, and since it's too tedious to dedicate time to managing what should and shouldn't be committed, I've created my own repository, mintlib-rc and freemint-rc (for release candidates).

Of course I remain available to help you guys if needed.